refactor: move writer and compactor permits into LeveledMap
# Summary

The writer and compactor semaphores, and the acquirer types that hand out
their permits, now live on `LeveledMap` instead of `SMV003`, removing the
dependency cycle between `leveled_store` and `sm_v003`.

# Details

`LeveledMap` already required both permits to freeze the writable level, so it
imported the permit types from `sm_v003`, while `sm_v003` owned the semaphores
and depended on `leveled_store` for everything else. Since the permits exist
to protect `LeveledMap`'s own level invariants, the map is where they belong:
the thing being protected now owns its locks.

`SMV003` keeps `acquire_writer_permit` and `acquire_compaction_and_writer` as
forwarding methods, so callers in the service crate are unaffected.

Snapshot installation is the one subtle case. It swaps in a fresh map but must
reuse the same semaphore objects, or a writer holding a permit from the
replaced map would stop excluding writers of the new one. That transfer used
to happen by copying two fields during `SMV003` construction; it is now
explicit in `LeveledMap::install_persisted`, which replaces the
`from_persisted` constructor. `test_install_snapshot_v003_shares_semaphores`
covers it and was confirmed to fail when the semaphores are not shared.
