+++ crates/server/raft-store/src/applier/applier_data/mod.rs
@@ -16,10 +16,10 @@ use std::sync::Arc;
 use std::sync::Mutex;
 use std::time::Duration;
 
+use crate::leveled_store::leveled_map::WriterPermit;
 use crate::leveled_store::state_machine::view::StateMachineView;
 use crate::sm_v003::Change;
 use crate::sm_v003::OnChange;
-use crate::sm_v003::writer_acquirer::WriterPermit;
 
 pub(crate) struct ApplierData {
     /// Hold a unique permit to serialize all apply operations to the state machine.
